Move substring sorting to a separate file, change the algorithm
– Instead of counting total substring characters, counting is now
done only up to branching point, i.e. to the first character that makes
the current substring suffix unique (without taking the separator
character into account). This reduces the number of required suffix
link operation when trying to find matches.
– Sorting uses backward_search when possible instead of just LF in
order to handle ranges of substrings with less operations.
– To make the code more readable, moved separate functions into classes
but inlined everything. This hopefully makes the code faster as
pushing parameters and return addresses to the stack is not needed.